Update vpg.py
The modification is in line 295. The original one is 
if timeout or epoch_ended:
However, it does not consider the situation that the game falls in the max_ep_len step. Therefore, the value of the final state s will be assigned with a non-zero scalar. We should classify this situation into the else branch. So change the code to
if (timeout or epoch_ended) and not d:
to ensure that the episode is interrupted AND the current state is not the final state.